+++ modules/module-mssql/src/replication/CDCReplicationJob.ts
@@ -88,7 +88,7 @@ export class CDCReplicationJob extends replication.AbstractReplicationJob {
     }
   }
 
-  async getReplicationLagMillis(): Promise<number | undefined> {
+  getReplicationLagMillis(): number | undefined {
     return this.lastStream?.getReplicationLagMillis();
   }
 }