+++ packages/service-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @powersync/service-core
 
+## 1.20.4
+
+### Patch Changes
+
+- df451c6: Node 24.14.0 and other dependency upgrades.
+- 11b4deb: Restructure `powersync_replication_lag_seconds` metric.
+- Updated dependencies [dea1e00]
+- Updated dependencies [ada86f2]
+  - @powersync/service-sync-rules@0.34.1
+  - @powersync/lib-services-framework@0.9.2
+  - @powersync/service-rsocket-router@0.2.18
+
 ## 1.20.3
 
 ## 1.20.2